Revert "Max avery/catalog validation"
