+++ src/app/admin/data-table.tsx
@@ -10,7 +10,7 @@ import {
     useReactTable,
 } from "@tanstack/react-table";
 
-import { approveUgcAdminAction as approveUgcAdmin } from "@/app/actions/serverActions";
+import { approveUgcAdmin } from "@/server/utils/queriesTS";
 
 import { Button } from "@/components/ui/button";
