+++ src/app/artist/[id]/_components/BlurbSection.tsx
@@ -2,7 +2,6 @@
 
 import { useState, useEffect } from "react";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
-import { aiPrompts, artists } from "@/server/db/schema";
 import { exists } from "drizzle-orm";
 
 
@@ -25,8 +24,6 @@ export default function BlurbSection({
   const [openModal, setOpenModal] = useState<'wiki' | 'ai' | null>(null);
   const [aiBlurb, setAiBlurb] = useState<string | undefined>();
   const [loadingAi, setLoadingAi] = useState(false);
-
-
 
   useEffect(() => {
     if (activeTab === "ai-generated" && !aiBlurb && !loadingAi) {
