+++ src/app/ugcstats/Leaderboard.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useEffect, useState } from "react";
-import type { LeaderboardEntry } from "@/app/actions/serverActions";
+import { LeaderboardEntry } from "@/server/utils/queriesTS";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { Badge } from "@/components/ui/badge";
 import { cn } from "@/lib/utils";
