+++ src/app/ugcstats/page.tsx
@@ -1,7 +1,7 @@
 import { getServerAuthSession } from "@/server/auth";
 import Dashboard from "./Dashboard";
 import { notFound } from "next/navigation";
-import { getUserById } from "@/server/utils/queries/userQueries";
+import { getUserById } from "@/server/utils/queriesTS";
 import Login from "../_components/nav/components/Login";
 import PleaseLoginPage from "../_components/PleaseLoginPage";
 
@@ -27,22 +27,7 @@ export default async function Page() {
     }
     
     // Normal authentication flow
-    if (!session) {
-        // Show dashboard in read-only mode for unauthenticated visitors
-        const guestUser = {
-            id: '00000000-0000-0000-0000-000000000000',
-            wallet: '0x0000000000000000000000000000000000000000',
-            email: null,
-            username: 'Guest User',
-            isAdmin: false,
-            isWhiteListed: false,
-            createdAt: new Date().toISOString(),
-            updatedAt: new Date().toISOString(),
-            legacyId: null
-        } as const;
-        return <Dashboard user={guestUser} />;
-    }
-
+    if (!session) return <PleaseLoginPage text="Login to view UGC Stats" />;
     const user = await getUserById(session.user.id);
     if (!user) return notFound();
     return <Dashboard user={user} />;
