--- src/server/utils/queries/index.ts
@@ -1,4 +1,0 @@
-export * from "@/server/utils/queries/artistQueries";
-export * from "@/server/utils/queries/userQueries";
-export * from "@/server/utils/queries/leaderboardQueries";
-export * from "@/server/utils/queries/discord";
