+++ src/server/utils/services.ts
@@ -1,7 +1,6 @@
-import { Artist, UrlMap } from "../db/DbTypes";
+import { Artist } from "../db/DbTypes";
 
-// Import directly from the artist queries module to ensure the symbol is recognised by TypeScript’s type checker
-import { getAllLinks } from "./queriesTS"; // Wrapper maintains compatibility with existing mocks
+import { getAllLinks } from "./queriesTS";
 
 export const artistWeb3Platforms = ['catalog', 'soundxyz', 'opensea', 'zora', 'mintsongs', 'supercollector', 'wallets', 'ens'];
 export const artistPlatforms = ['catalog', 'soundxyz', 'opensea', 'zora', 'mintsongs', 'x', 'audius', 'bandisintown', 'ens', 'wallets', 'facebook', 'instagram', 'lastfm', 'soundcloud', 'tiktok', 'youtubechannel', 'supercollector'];
@@ -35,11 +34,18 @@ type SpotifyDataType = {
 }
 
 export const getArtistDetailsText = (artist: Artist, spotifyData: SpotifyDataType) => {
+    let web3Platforms = getArtistSplitPlatforms(artist).web3Platforms
     const numSpotifyReleases = (spotifyData != null && spotifyData.releases != null) ? spotifyData.releases : 0;
-    if (numSpotifyReleases <= 0) return "";
+    if (web3Platforms.length <= 0 && numSpotifyReleases <= 0) return "";
 
-    if (numSpotifyReleases > 0) return `${numSpotifyReleases} releases on Spotify`;
-    return `${numSpotifyReleases} releases on Spotify`;
+    if (web3Platforms.length <= 0 && numSpotifyReleases > 0) return `${numSpotifyReleases} releases on Spotify`;
+
+    const prefix = numSpotifyReleases > 0 ? `${numSpotifyReleases} releases on Spotify; NFTs released on ` : "NFTs released on "
+    if (web3Platforms.length < 2) return prefix + web3Platforms[0]
+
+    if (web3Platforms.length > 1)
+        web3Platforms[-1] = `and ${web3Platforms[-1]}`
+    return prefix + web3Platforms.join(", ")
 }
 
 export function isObjKey<T extends object>(key: PropertyKey, obj: T): key is keyof T {
@@ -150,7 +156,7 @@ export async function extractArtistId(artistUrl: string) {
     }
 
     // Fallback for SoundCloud username URLs not caught by DB regex
-    const soundCloudRow = allLinks.find((l: UrlMap) => l.siteName === 'soundcloud');
+    const soundCloudRow = allLinks.find(l => l.siteName === 'soundcloud');
     if (soundCloudRow && artistUrl.includes('soundcloud.com')) {
         try {
             const url = new URL(artistUrl.startsWith('http') ? artistUrl : `https://${artistUrl}`);
